+++ UnityProject/Assets/Prefabs/Items/Construction/Assemblies/RemoteSignalingDevice.prefab
@@ -120,7 +120,7 @@ PrefabInstance:
         type: 3}
       propertyPath: customNetTransform
       value: 
-      objectReference: {fileID: 7044771064810597979}
+      objectReference: {fileID: 0}
     - target: {fileID: 7396244808014028178, guid: 201d4b5714989d64d86b11da761ecb15,
         type: 3}
       propertyPath: foreverID
@@ -154,7 +154,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   syncMode: 0
   syncInterval: 0.1
-  signalData: {fileID: 11400000, guid: ac9bcd63f2f8a944ebd2001ce5c0a3b1, type: 2}
+  emmitableSignalData:
+  - {fileID: 11400000, guid: ac9bcd63f2f8a944ebd2001ce5c0a3b1, type: 2}
   passCode: 0
   frequency: 122
   requiresPower: 0
@@ -174,15 +175,3 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   NetTabType: 48
   requiresAltClick: 1
---- !u!114 &7044771064810597979 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 4578942373061721135, guid: 201d4b5714989d64d86b11da761ecb15,
-    type: 3}
-  m_PrefabInstance: {fileID: 6795833047291761268}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6940507403175904139}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 579884b05ae74e8499329b64997147d3, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
