+++ UnityProject/Assets/Scripts/Messages/Server/ServerTypingMessage.cs
@@ -37,6 +37,11 @@ namespace Messages.Server
 			var showTyping = msg.state == TypingState.TYPING;
 
 			// check if player is conscious before generating typing icon
+			if (player.playerHealth == null)
+			{
+				icon.ToggleTypingIcon(showTyping);
+				return;
+			}
 			bool isPlayerConscious = (player.playerHealth.ConsciousState == ConsciousState.CONSCIOUS ||
 			                          player.playerHealth.ConsciousState == ConsciousState.BARELY_CONSCIOUS);
 			if (isPlayerConscious)
