--- UnityProject/Assets/Scripts/Objects/Wallmounts/PublicDepartmentTerminal.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9b765ac9b30643c08432e68ddb18dd08
-timeCreated: 1648935292
