--- UnityProject/Assets/Scripts/Objects/Wallmounts/PublicTerminalReceiver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d2571f1f49edba47abeb2caf877b73e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
