+++ UnityProject/Assets/Scripts/Objects/Wallmounts/PublicTerminals/PublicTerminalModule.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace Objects.Wallmounts.PublicTerminals
+{
+	public class PublicTerminalModule : MonoBehaviour
+	{
+		public bool isActive = true;
+		public PublicDepartmentTerminal Terminal;
+	}
+}
